Adding text editor for writing new recipes
Using draftJS as a text editor. Recipe text is converted to markdown and sent off to the database.
using markdown as it is more lightweight and can be used in places more than just a browser page.

This also required updating the tsconfig to enable css modules to be loaded in typescript.
This is why there is now a global.d.ts that declares the css/less files etc.

Also fixed the URL endpoint for the PUT call to the backend.