Recipes can now have tags
The CollectionTable stores the list of tags in a separate table joining by the recipe id.
Will look to see if there is a way to avoid storing duplicate values, ie: recipe 1 and 2 both
have tag 'food' then there will be 2 rows. Probably a pointless optimization though.

Client is updated to show tags and to include tags in the PUT request