Tweaking HTMLProcessor
Now we only pull out the text from the HTML. May later look at separating instructions&ingredients into different elements
in the list. But given the end user will be editing and formatting this it's not 100% necessary right now.
Basic heuristic seems pretty effective with a few example sites. There are some minor issues like picking up the instructions in the
ingredients section also. But again given the end user will edit this before save it's not a big deal